spi:sunxi: fixes a bug with setting up spi transfers
This patch addresses an issue where the first transfers to an spi device
doesn't get set up correctly if it just uses the parameters specified in
the script.bin file and has no overrides in the transfer structure.  The
patch forces the parameters to be setup at the start of each group of
transfers and also removes unnecessary initialisation in spi_sunxi_hw_init.

Signed-off-by: Simon Ratcliffe simon@proatomic.com.au